requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
ebaysdk==2.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.13.1
gunicorn==21.2.0
